Create_syo/requirements.txt
pandas
openpyxl
streamlit
re
streamlit_extras.grid
decouple
sqlalchemy
